move scripts into a separate folder
