+++ src/scripts/attack_commend.js
