+++ src/scripts/attack_dismental.js
