+++ src/scripts/attack_yitiji.js
