+++ src/scripts/script_claimNewRoom.js
