+++ src/scripts/script_invaderCore.js
