+++ src/scripts/script_outerEnergyBase_byRoom.js
