+++ src/scripts/script_supportEnergy.js
